cms/pyproject.toml
[project]
name = "cms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "img2pdf",
    "ocrmypdf",
    "ollama",
    "pillow",
    "pypdf",
    "pydantic",
    "pymupdf",
    "pyside6",
]

